update download button and add the animation

+++ app/src/main/java/org/nekomanga/domain/chapter/Chapter.kt
@@ -120,7 +120,7 @@ data class SimpleChapter(
 
     }
 
-    fun toChapterItem(downloadState: Download.State = Download.State.default, downloadProgress: Int = 0): ChapterItem = ChapterItem(
+    fun toChapterItem(downloadState: Download.State = Download.State.default, downloadProgress: Float = 0f): ChapterItem = ChapterItem(
         chapter = this,
         downloadState = downloadState,
         downloadProgress = downloadProgress,
@@ -155,7 +155,7 @@ fun Chapter.toSimpleChapter(): SimpleChapter? {
 data class ChapterItem(
     val chapter: SimpleChapter,
     val downloadState: Download.State = Download.State.default,
-    val downloadProgress: Int = -1,
+    val downloadProgress: Float = -1f,
 ) {
     val isDownloaded = downloadState == Download.State.DOWNLOADED
 }
